Adds hdi metadata, and Social contextual type to contextual layer category enum and file rename
